% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{load_study_x}
\alias{load_study_x}
\title{load all vector and raster files of study a in package AniMove}
\usage{
load_study_x(overwrite = TRUE)
}
\description{
All files are spatial vector data sets (Polygons, Lines or Points) providing information
}
\details{
the files are

"lsat_extent_LL": landsat extent for path 224 r63 (geographic coordinates lat long)

"lsat_extent": landsat extent for path 224 r63 (projection UTM north WGS84)

"study_area": rectangle of the study area (projection UTM north WGS84)
}
\examples{
load_vector()
## plotting the data in geographic coordinates
plot(lsat_extent_LL)

## plotting the data in UTM north projection
plot(pa)
plot(abspres_poi,col="yellow",add=T)
}
\author{
... Martin Wegmann ....
}
\references{
\href{http://www.animove.org}.
}
\seealso{
load_study_()
}
\keyword{data}
\keyword{frames}
\keyword{shape}
\keyword{spatial}
\keyword{vector,}

